move tests into _core/_tests/test_run.py
